chore: remove web3auth and db
