+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,29 +1,5 @@
 import NextAuth from "next-auth";
-import GoogleProvider from "next-auth/providers/google";
-import type { AuthOptions } from "next-auth";
-
-export const authOptions: AuthOptions = {
-  secret: process.env.AUTH_SECRET,
-  providers: [
-    GoogleProvider({
-      clientId: process.env.GOOGLE_CLIENT_ID!,
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
-    }),
-  ],
-  callbacks: {
-    async jwt({ token, user, account }) {
-      if (account) {
-        token.id_token = account.id_token;
-      }
-      return token;
-    },
-
-    async session({ session, token }) {
-      session.id_token = token.id_token;
-      return session;
-    },
-  },
-};
+import { authOptions } from "@/lib/authOptions";
 
 const handler = NextAuth(authOptions);
