+++ src/lib/authOptions.ts
@@ -1,0 +1,25 @@
+import GoogleProvider from "next-auth/providers/google";
+import type { AuthOptions } from "next-auth";
+
+export const authOptions: AuthOptions = {
+  secret: process.env.AUTH_SECRET,
+  providers: [
+    GoogleProvider({
+      clientId: process.env.GOOGLE_CLIENT_ID!,
+      clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
+    }),
+  ],
+  callbacks: {
+    async jwt({ token, user, account }) {
+      if (account) {
+        token.id_token = account.id_token;
+      }
+      return token;
+    },
+
+    async session({ session, token }) {
+      session.id_token = token.id_token as string;
+      return session;
+    },
+  },
+};
